Function tree_magic::is_alias

pub fn is_alias(mime1: String, mime2: String) -> bool

Determines if a MIME is an alias of another MIME

If this returns true, that means the two MIME types are equivalent. If this returns false, either one of the MIME types are missing, or they are different. If you're using the staticmime feature flag, input is a &'static str. Otherwise it is a String.

Examples

let mime1 = "application/zip".to_string();
let mime2 = "application/x-zip-compressed".to_string();

assert_eq!( tree_magic::is_alias(mime1, mime2), true );